+++ pcs/usage.py
@@ -516,8 +516,10 @@ Commands:
         resource will be automatically removed once the resource is running on
         it's new location. The command will fail in case it is not possible to
         verify that the resource will not be moved after deleting the
-        constraint. If --strict is specified, the command will also fail if
-        other resources would be affected.
+        constraint. If this happens after the location constraint has been
+        created, the constraint will be left in the configuration. If --strict
+        is specified, the command will also fail if other resources would be
+        affected.
 
         If --wait is specified, pcs will wait up to 'n' seconds for the
         resource to move and then return 0 on success or 1 on error. If 'n' is